pinecone-rag-app/requirements.txt
langchain
langchain-google-genai
langchain-pinecone
pinecone-client
python-dotenv
langchain-community
langchain-text-splitters
